Moved yaml-migrate from exp to cmd
